requirements.txt
cv2
aio-pika==6.4.1
aiohttp==3.6.2
aiormq==3.2.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
multidict==4.7.2
pamqp==2.3.0
pika==1.1.0
PyMySQL==0.9.3
pynsq==0.9.0b1
redis==3.3.11
requests==2.22.0
SQLAlchemy==1.3.12
tornado==4.5.3
urllib3==1.25.7
yarl==1.4.2
